models, utils, scraper done
