+++ src/mcf_rag_chatbot/backend/api.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
-from mcf_rag_chatbot.backend.data_models import Prompt
-from mcf_rag_chatbot.backend.rag import rag_agent
+from .data_models import Prompt
+from .rag import rag_agent
 app = FastAPI()
 
 # root page message
