+++ src/mcf_rag_chatbot/webscraper/__init__.py
@@ -1,0 +1,2 @@
+def main() -> None:
+    print("Hello from ragbit!")
