+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka.json
@@ -1,0 +1,47 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/",
+  "title": "Hemberedskap",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Betalningar och kontanter"
+    },
+    {
+      "level": "h2",
+      "text": "Grunder i hemberedskap"
+    },
+    {
+      "level": "h2",
+      "text": "Extra stöd"
+    },
+    {
+      "level": "h2",
+      "text": "Mat"
+    },
+    {
+      "level": "h2",
+      "text": "Värme"
+    },
+    {
+      "level": "h2",
+      "text": "Vatten"
+    },
+    {
+      "level": "h2",
+      "text": "Kommunikation"
+    },
+    {
+      "level": "h2",
+      "text": "Medicin för en månad hemma"
+    },
+    {
+      "level": "h2",
+      "text": "Beredskap för dina husdjur"
+    },
+    {
+      "level": "h2",
+      "text": "Kurser och föreläsningar om hemberedskap"
+    }
+  ],
+  "content": "Att ha hemberedskap innebär att kunna klara sig utan samhällets hjälp i minst en vecka. Om du kan preppa för en vecka och ta ansvar för dig själv och din familj, så bidrar du till att hålla Sveriges gemensamma beredskap på en hög nivå.\nAtt ha hemberedskap innebär att kunna klara sig utan samhällets hjälp i minst en vecka. Om du kan preppa för en vecka och ta ansvar för dig själv och din familj, så bidrar du till att hålla Sveriges gemensamma beredskap på en hög nivå.\nI kris och krig kan det bli svårare att köpa mat, medicin och drivmedel. Se därför till att kunna betala på flera olika sätt.\nGör en plan för hur du kan ordna med vatten, mat, kommunikation och värme hemma under minst sju dagar.\n- För dig med nedsatt syn\n- För dig med nedsatt hörsel\n- För dig som är rörelsehindrad\n- För dig med  kognitiva svårigheter\n- Odla själv\n- Tillgång till mat i kris\nDin bostad blir snabbt kall om det är strömavbrott på vintern.\n- Tillgång till vatten i kris\n- Tvätt och toalett\nVid en större olycka eller en kris behöver du få veta vad som händer, hur du själv ska agera och vad ansvariga myndigheter gör.\nOm du brukar få medicin eller förbrukningsartiklar utskrivna av sjukvården är det bra att ha hemma så att det räcker för en månad.\nDu som har husdjur är ansvarig för att djuret mår bra både i vardagen och vid kris eller krig.\nDu som vill ha en föreläsning eller kurs i hemberedskap kan hitta den här."
+}
