+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_beredskap-for-dina-husdjur.json
@@ -1,0 +1,15 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/beredskap-for-dina-husdjur/",
+  "title": "Beredskap för dina husdjur",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Evakuering och skyddsrum"
+    },
+    {
+      "level": "h2",
+      "text": "Checklistor för husdjuren"
+    }
+  ],
+  "content": "Du som har husdjur är ansvarig för att djuret mår bra både i vardagen och vid kris eller krig. Tänk igenom vilka behov just ditt eller dina husdjur har och skaffa det som behövs.\nDet är viktigt att du tänkt igenom hur du bäst tar hand om ditt djur om en kris inträffar. Om till exempel strömmen försvinner så ska ni kunna ha tillgång till:\n- Vatten\n- foder\n- mediciner\n- värme.\nHa ett lager hemma för minst en vecka. Använd alltid det äldsta fodret först så att det inte blir gammalt.\nOm du behöver evakueras från din bostad till exempel vid en skogsbrand eller ett farligt utsläpp, ta med ditt husdjur. Ha en bur, väska eller något annat att transportera djuret i. Träna gärna djuret att vara däri, så att den blir enkel att använda.\nHusdjur ska inte tas med till skyddsrum. Uppsök istället annat skyddat utrymme som till exempel källare, garage eller en tunnelbanestation. Måste du lämna husdjuren hemma kan du ställa fram extra foder och vatten, så har djuret goda förutsättningar att klara sig på egen hand tills du är tillbaka.\nOrganisationen Svenska Blå Stjärnan har samhällets uppdrag att ta hand om djuren vid kriser, katastrofer, utbrott av smitta och andra händelser som kan drabba djuren. De har tagit fram checklistor som gäller för våra vanligaste husdjur, det vill säga katter, hundar, fåglar, reptiler, gnagare, kanin och häst.\nMed listorna kan du planera för hur du löser det med värme, vatten, hygien och foder. Sätt upp dem på ett ställe där alla i hushållet vet att de alltid finns.\nSvenska Blå Stjärnan har också gjort ett utbildningsmaterial för barn som handlar om vad en samhällskris är och hur den kan påverka samhället, människorna och djuren."
+}
