+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_betalningar-och-kontanter-i-kris.json
@@ -1,0 +1,31 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/betalningar-och-kontanter-i-kris/",
+  "title": "Betalningar och kontanter i kris",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Så förbereder du dig"
+    },
+    {
+      "level": "h2",
+      "text": "Att skaffa kontanter"
+    },
+    {
+      "level": "h2",
+      "text": "Se till att kunna betala på flera olika sätt"
+    },
+    {
+      "level": "h2",
+      "text": "Fördelen med kontanter"
+    },
+    {
+      "level": "h2",
+      "text": "Riksbankens roll"
+    },
+    {
+      "level": "h2",
+      "text": "Frågor och svar om betalningar och kontanter"
+    }
+  ],
+  "content": "I kris och krig kan det bli svårare att köpa mat, medicin och drivmedel. Se därför till att kunna betala på flera olika sätt. Myndigheten för civilt försvar och Riksbanken rekommenderar att du har kontanter hemma för minst en vecka.\nFoto: Johnér\nSe för det första till att du har mat och annat nödvändigt så att du klarar dig minst en vecka utan att handla. På det sättet minskar du behovet av att behöva betala vid en kris. Diskutera med familj, vänner eller grannar hur ni skulle göra om ni inte kan köpa till exempel mat, mediciner eller drivmedel. Räkna på kostnaden för din familj under minst en veckas tid och ha kontanter hemma i små valörer att handla med om digitala betalsätt inte fungerar.\nFundera över:\n- Vilken hemberedskap har ni i form av vattendunkar, mat med lång hållbarhet som inte behöver tillagas och till exempel mediciner?\n- Vad saknas hemma som kan komma att behöva handlas under den första veckan?\n- Vilka alternativ har ni om er bank eller er betalkortsleverantör (till exempel VISA eller Mastercard) drabbas av problem som gör det svårt för er att betala?\n- Vilka alternativ har ni om er vanliga butik drabbas av problem som gör det svårt för er att betala? Finns det andra butiker där ni kan handla och kan ni ta er dit?\n- Vad har ni för andra möjligheter att hjälpa andra och vilka i din närhet kan hjälpa dig?\nIdag sker de flesta betalningar i Sverige sedan länge med kort. Kontanterna är inte lika synliga i samhället som de var förr. Men det finns gott om kontanter i systemet. Här är några sätt att få tag i dem.\n- Via uttagsautomter, till exempel Bankomat, ICA-banken och Kontanten.\n- När du betalar med kort i en butik kan du be om att få göra en avrundning uppåt på ditt köp och få ut mellanskillnaden i kontanter. Det förutsätter att butiken hanterar kontanter. Du behöver förstås ha täckning på kortet.\n- Du kan beställa kontanter i valfri valuta från växlingskontor och få dem levererade till ditt postombud.\n- Du kan ta ut pengar på växlingskontor.\n- Som kund i en bank kan du i flera fall beställa svenska och utländska sedlar i internetbanken för privatpersoner. Pengarna skickas rekommenderat till ditt postombud. Fråga din bank.\nEn viktig beredskapsåtgärd är att ha flera sätt att betala. I ditt hushåll kan ni till exempel:\n- Ha konton i olika banker, kort från olika kortnätverk (exempelvis Visa och Mastercard) och tillgång till betaltjänster via mobiltelefon (exempelvis Swish).\n- Ha flera alternativ till internetuppkoppling.\n- Ha flera alternativ att identifiera dig elektroniskt (exempelvis BankID och säkerhetsdosa till din bank).\n- Ha kontanter i mindre valörer som räcker för en veckas inköp för ditt hushåll.\n- Använd kontanter då och då för att upprätthålla din beredskap och vara van vid hur kontanter ser ut och fungerar.\n- Ha koll på era betalkort och PIN-koder om kontaktlösa betalningar med kort och mobil (genom att ”blippa” mot en kortterminal) skulle sluta fungera.\nÄven om andra betalningssätt har blivit mycket vanligare, så har kontanter fortfarande en viktig roll i samhället.\n- Det är det enda betalningsmedel som alla kan ha tillgång till.\n- Du behöver inte någon speciell teknik som appar, betalkort eller e-legitimation för att betala.\n- Du kan ta ut och förvara en summa hemma i beredskap.\n- Kontanter är centralbankspengar, det vill säga att värdet garanteras av staten.\n- Det krävs inga mellanhänder för att genomföra betalningar med kontanter.\nPengar och betalningar fyller en viktig funktion i samhället och Riksbankens mål är att betalningar i Sverige ska vara säkra, effektiva och tillgängliga. Riksbanken har ett särskilt ansvar för att allmänheten kan genomföra betalningar i fredstida krissituationer och vid höjd beredskap.Stora delar av betalningsverksamheten drivs av banker och andra privata företag. Enligt riksbankslagen får Riksbanken ställa krav på att företag som är viktiga för betalningar ska planera för att kunna upprätthålla sin betalningsverksamhet även vid kriser.Innehållet på denna sida är framtaget tillsammans med Riksbanken."
+}
