+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_extra-stod.json
@@ -1,0 +1,23 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/extra-stod/",
+  "title": "För dig som behöver extra stöd i kris och krig",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "För dig med nedsatt syn"
+    },
+    {
+      "level": "h2",
+      "text": "För dig med nedsatt hörsel"
+    },
+    {
+      "level": "h2",
+      "text": "För dig som är rörelsehindrad"
+    },
+    {
+      "level": "h2",
+      "text": "För dig med  kognitiva svårigheter"
+    }
+  ],
+  "content": "De här sidorna innehåller beredskapstips för dig som har nedsatt förmåga och ofta behöver extra stöd i vardagen. Du kommer också att behöva det i en kris eller ett krig. Grannar, anhöriga och assistenter får också gärna ta del av tipsen. Vi behöver alla hjälpas åt.\nDe här sidorna innehåller beredskapstips för dig som har nedsatt förmåga och ofta behöver extra stöd i vardagen. Du kommer också att behöva det i en kris eller ett krig. Grannar, anhöriga och assistenter får också gärna ta del av tipsen. Vi behöver alla hjälpas åt.\nTips om beredskap för kriser och krig. För dig som är blind eller har nedsatt syn.\nTips om beredskap för kriser och krig. För dig som är döv eller hörselskadad.\nTips om beredskap för kriser och krig. För dig som är rörelsehindrad.\nTips om att vara förberedd för kriser och krig. För dig med kognitiv funktionsnedsättning."
+}
