+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_extra-stod_blinda-och-synnedsattning.json
@@ -1,0 +1,23 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/extra-stod/blinda-och-synnedsattning/",
+  "title": "Beredskap för dig med nedsatt syn",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Några steg att börja med"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista för dig med synnedsättning"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista för dig som hjälper blinda och synnedsatta"
+    },
+    {
+      "level": "h2",
+      "text": "Vi tar emot fler idéer!"
+    }
+  ],
+  "content": "Tips om beredskap för kriser och krig, för dig som är blind eller har nedsatt syn. Ju mer kunskap du har, desto mer förberedd och trygg blir du.\nVid en kris kan en synnedsättning innebära nya utmaningar, särskilt om du har svårt att orientera och förflytta dig i ovana miljöer eller ta del av skriftlig information.\nNär vardagen inte fungerar som vanligt är det inte säkert att hjälp automatiskt hittar fram till dig. Du behöver därför själv vara beredd att ta kontakt med andra för att få hjälp att hitta vägen för att hämta rent dricksvatten, ta dig till en kommunal samlingsplats eller för att läsa uppsatta texter och skyltar.\n- Kontakta din kommun och fråga vad som gäller för dig vid en kris. Prata med beredskapsansvarig eller din biståndshandläggare.\n- Prata gärna med grannar och dina närstående om hur ni kan hjälpas åt. Det är bra om personer i din omgivning vet om att du har en synnedsättning.\n- Förbered hur du kan ta med dig viktiga hjälpmedel vid vid en evakuering eller en situation då du måste lämna ditt hem.\n- Du också ta kontakt med Synskadades Riksförbund, SRF. De har egna krisinstruktörer som själva har en synnedsättning.Telefon: 08-39 90 00E-post:info@srf.nu\nÖva för att klara krisen bättre\nÖva gärna på en påhittad krissituation, exempelvis ett elavbrott. Om möjligt tillsammans med personer som är viktiga för dig om något sådant skulle hända. Pröva att ordna med mat, få tag på viktig information,  kommunicera med andra och sköta dina dagliga sysslor. Genom att öva på att göra sådant utan att använda el kan du upptäcka brister i din beredskap som du inte tänkt på tidigare.\n- Bekanta dig med och öva på olika utrymningsvägar från ditt boende och/eller din arbetsplats.\n- Ta reda på om det finns skyddsrum eller trygghetspunkter där du bor eller arbetar. Försök få möjlighet att öva på att ta dig dit. Ledarhund och assistanshund får följa med till skyddsrum.\nDet är bra att åtgärda bristerna innan en verklig kris inträffar. Övning gör dig mer medveten och stärker din förmåga att hantera en kris!"
+}
