+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_extra-stod_dova-och-horselskadade.json
@@ -1,0 +1,27 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/extra-stod/dova-och-horselskadade/",
+  "title": "Beredskap för dig med nedsatt hörsel",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Krislåda för dig med hörselnedsättning eller dövhet"
+    },
+    {
+      "level": "h2",
+      "text": "Öva för att klara krisen bättre"
+    },
+    {
+      "level": "h2",
+      "text": "Råd till dig som är döv eller hörselskadad"
+    },
+    {
+      "level": "h2",
+      "text": "För dig som samarbetar med döva eller hörselskadade personer i kris eller krig"
+    },
+    {
+      "level": "h2",
+      "text": "Vi tar emot fler idéer!"
+    }
+  ],
+  "content": "Tips om beredskap för kriser och krig, för dig som är döv eller hörselskadad. Ju mer kunskap du har, desto mer förberedd och trygg blir du.\nVid en kris kan en hörselskada innebära nya utmaningar, särskilt om du har svårt att uppfatta larm och meddelanden och att kommunicera med andra.\nDet är inte säkert att hjälp hittar fram till dig automatiskt. Du behöver därför själv vara beredd att ta kontakt med andra för att få hjälp att hitta vägen för att hämta rent dricksvatten, ta dig till en kommunal samlingsplats eller för att läsa uppsatta texter och skyltar.\nHa vardaglig kontakt med dina grannar, din bostadsrättsförening, hyresvärd eller samfällighet så att de som bor nära är medvetna om att du kan behöva hjälp vid en kris. Se till att ni har varandras kontaktuppgifter. De kan exempelvis förse dig med information du behöver, eller meddela dig om ni behöver uppsöka skyddsrum.Ta gärna kontakt med din kommun och undersök hur deras planer och beredskap ser ut för döva och hörselskadade. Ju mer kunskap du har desto mer förberedd och tryggare blir du!\nHa en krislåda med saker som hjälper dig att klara din vardag i minst en vecka vid kris och krig. Det finns saker som är nödvändiga för alla, som mat, vatten och värme. Det finns många tips i broschyren \"Om krisen eller kriget kommer\" och här på mcf.se. Men du behöver också tänka igenom hur dina särskilda behov och förutsättningar ser ut som döv eller hörselskadad och anpassa innehållet i din egen krislåda efter det. Se till att flera i din omgivning vet var du har krislådan!\n- Nedskrivet på ett inplastat papper: Information om att du är döv eller hörselskadad och om hur man kan kommunicera med dig . Är du tolkanvändare, skriv telefonnumret till tolkcentralen och var man kan boka tolk utanför kontorstid. Skriv även vilken tolkmetod du önskar, exempelvis teckenspråkstolk eller skrivtolk.\n- Extra och uppladdade batterier till dina hjälpmedel. Har du laddningsbara hörapparater, se till att ha en laddad powerbank. Har du hjälpmedel som går på el, ta reda på hur de kan användas om strömmen gått. Det finns elverk för hemmabruk men du, dina anhöriga eller andra i din närhet behöver veta hur det ska användas.\n- Papper och penna för skriven kommunikation om dina tekniska hjälpmedel inte fungerar, eller om personer i din närhet inte kan teckenspråk, om du använder det.\nÖvning stärker din förmåga att hantera en kris och gör dig mer medveten. Öva en påhittad krissituation, till exempel ett strömavbrott, tillsammans med personer som är viktiga för dig. Pröva att ordna med mat, få tag på viktig information, kommunicera med anhöriga och vänner och sköta dina dagliga rutiner och behov utan att använda el. Genom att öva kan du upptäcka brister i din beredskap som du kanske inte tänkt på tidigare. Du har också möjlighet att åtgärda bristerna innan en verklig kris inträffar.\nTa gärna kontakt med din kommun och undersök hur deras krisberedskap ser ut för döva och hörselskadade. Ju mer kunskap du har desto mer förberedd och tryggare blir du."
+}
