+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_extra-stod_for-dig-med--if-intellektuell-funktionsnedsattning.json
@@ -1,0 +1,31 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/extra-stod/for-dig-med--if-intellektuell-funktionsnedsattning/",
+  "title": "Beredskap för dig med kognitiv funktionsnedsättning",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Hemberedskap"
+    },
+    {
+      "level": "h2",
+      "text": "Hjälp från andra"
+    },
+    {
+      "level": "h3",
+      "text": "Öva på att vara kvar hemma"
+    },
+    {
+      "level": "h3",
+      "text": "Öva på att lämna hemmet"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista för dig som hjälper någon annan"
+    }
+  ],
+  "content": "De flesta i Sverige ska kunna klara sig själva minst en vecka i kris eller krig. Be någon om hjälp att förbereda dig. Du behöver till exempel ha vatten, mat, mediciner och batterier.\nHemberedskap betyder att du har förberett digför att klara din vardag minst en vecka.Vad använder du varje dag?De sakerna ska du ha även i kris och krig.Du behöver också kunna följa nyhetersom är tydligaför dig.\nOm du har personlig assistent eller får hjälp av personal på ditt boendekommer du att få den hjälpen även i kris eller krigmen det kan ta längre tid.Om det blir kris eller krig i Sverigeska kommunen och alla olika myndigheterse till att samhället fungerar som vanligt.Frivilliga organisationerkan också hjälpa till.Alla måste arbeta tillsammans.\nÖva på krisMan kan passa på att öva när allt är lugnt.När man övar så ser man direktvad man behöver bli bättre påtills något händer på riktigt.Det är bra att öva med flerasom en granne, kompis, familj eller assistent.\nLåtsas att det är strömavbrott flera timmar.Se hur du kan be någon om hjälp.Sköt dina dagliga sysslor utan ström.Prova att äta något till middag som inte behöver värmas upp.Lyssna på Sveriges Radio P4 under dagen.Radion behöver batterier.\nLåtsas att du måste gå hemifrån för att det är oväder eller krig.Se om det finns skyddsrumeller trygghetspunkternära där du bor.Öva på att ta dig dit.Om du är någon annanstans på dagarna,på skola eller jobb,så ska du veta var skyddsrummen finns nära där också."
+}
