+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_extra-stod_rorelsehindrad.json
@@ -1,0 +1,19 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/extra-stod/rorelsehindrad/",
+  "title": "Beredskap för dig som är rörelsehindrad",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Krislåda för dig med rörelsehinder"
+    },
+    {
+      "level": "h2",
+      "text": "Öva för att klara krisen bättre"
+    },
+    {
+      "level": "h2",
+      "text": "Vi tar emot fler idéer!"
+    }
+  ],
+  "content": "Tips om beredskap för kriser och krig, för dig som är rörelsehindrad. Ju mer kunskap du har, desto mer förberedd och trygg blir du.\nHa vardaglig kontakt med dina grannar, din bostadsrättsförening, hyresvärd eller samfällighet så att de som bor nära är medvetna om att du kan behöva hjälp vid en kris. Se till att ni har varandras kontaktuppgifter. Diskutera med anhöriga, vänner, grannar, assistenter eller biståndshandläggare om vilken hjälp du kan behöva.\nTa gärna kontakt med din kommun och undersök hur deras planer och beredskap ser ut för rörelsehindrade. Ju mer kunskap du har desto mer förberedd och tryggare blir du!\nHa en krislåda med saker som hjälper dig att klara din vardag i minst en vecka vid kris och krig. Det finns saker som är nödvändiga för alla, som mat, vatten och värme. Det finns många tips i broschyren \"Om krisen eller kriget kommer\" och här på mcf.se. Men du behöver också tänka igenom hur dina särskilda behov och förutsättningar ser ut som rörelsehindrad och anpassa innehållet i din egen krislåda efter det. Se till att flera i din omgivning vet var du har krislådan!\n- Om du använder elrullstol, ha gärna en lätt manuell stol som reserv.\nOm du använder elrullstol, ha gärna en lätt manuell stol som reserv.\n- Om du använder dyna, och du måste evakuera utan rullstol, ta dynan med dig.\nOm du använder dyna, och du måste evakuera utan rullstol, ta dynan med dig.\n- Ha en extra rörlighetsenhet som käpp eller rullator hemma.\nHa en extra rörlighetsenhet som käpp eller rullator hemma.\n- Ha verktyg som kan behövas för din utrustning, till exempel skiftnyckel.\nHa verktyg som kan behövas för din utrustning, till exempel skiftnyckel.\n- Skaffa en pannlampa, de är bra för att frigöra händerna.\nSkaffa en pannlampa, de är bra för att frigöra händerna.\n- Ha en visselpipa för att kunna signalera efter hjälp.\nHa en visselpipa för att kunna signalera efter hjälp.\nÖvning stärker din förmåga att hantera en kris och gör dig mer medveten. Öva en påhittad krissituation, till exempel ett strömavbrott, tillsammans med personer som är viktiga för dig. Pröva att ordna med mat, få tag på viktig information, kommunicera med anhöriga och vänner och sköta dina dagliga rutiner och behov utan att använda el. Genom att öva kan du upptäcka brister i din beredskap som du kanske inte tänkt på tidigare. Du har också möjlighet att åtgärda bristerna innan en verklig kris inträffar.\nTa gärna kontakt med din kommun och undersök hur deras krisberedskap ser ut för rörelsehindrade. Ju mer kunskap du har desto mer förberedd och tryggare blir du.\nDu kan också gå en utbildning i hemberedskap, som bland annat Intressegruppen för Assistansberättigade anordnar.\nTill Intressegruppen för Assistansberättigade (IFA).\nDe här tipsen är framtagna av Myndigheten för civilt försvar tillsammans med personer som själva har någon funktionsnedsättning eller ett särskilt behov av stöd.\nVälkommen att kontakta oss om du saknar något eller vill dela med dig av fler tips och råd för en bättre beredskap.\nTelefonnummer 077-124 02 40 ellerkontakt@mcf.se."
+}
