+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_grunder-i-hemberedskap.json
@@ -1,0 +1,31 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/grunder-i-hemberedskap/",
+  "title": "Grunder i hemberedskap",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Se till att du är förberedd"
+    },
+    {
+      "level": "h2",
+      "text": "Utmana dig! Klarar du krisen i sju dagar?"
+    },
+    {
+      "level": "h2",
+      "text": "Ta hand om dig själv"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista"
+    },
+    {
+      "level": "h2",
+      "text": "Om du känner dig orolig"
+    },
+    {
+      "level": "h2",
+      "text": "Frågor och svar om hemberedskap"
+    }
+  ],
+  "content": "Hemberedskap betyder att du har det du och dina närstående behöver hemma för att klara er i sju dagar. Här finns tips på vad du kan ha i en krislåda för att bättre kunna hantera strömavbrott, extremväder eller andra kriser.\nFoto:Johnér\nVårt samhälle är sårbart. Inte minst för att vi behöver el och internet till väldigt mycket. En kris kan göra vardagen besvärlig på väldigt kort tid.\nDet här kan hända:\n- Värmen försvinner.\n- Det blir svårt att laga och förvara mat.\n- Mat och andra varor kan ta slut i affärerna.\n- Det kommer inget vatten i kranen eller toaletten.\n- Det går inte att tanka bilen.\n- Betalkort och bankomater fungerar inte.\n- Mobilnät och internet fungerar inte.\n- Kollektivtrafik och andra transporter står stilla.\n- Det blir svårt att få tag i läkemedel och medicinsk utrustning.\nAlla som kan, behöver därför ha en hemberedskap. Om så många som möjligt av oss klarar vår vardag utan samhällets service i minst en vecka, får samhället tid att samla ihop sina resurser för att kunna hantera situationen så bra som möjligt. Fokus kommer då att ligga på att hjälpa dem som har svårast att klara sig själva. Du som vanligtvis behöver stöd i vardagen har rätt att få den hjälpen även i en kris, men hjälpen kan se annorlunda ut, eller dröja.\nDu kan behöva anpassa våra råd efter just ditt hushålls behov och förutsättningar. Vad behöver du eller dina närstående under en vanlig vecka?\nVi lever i en tid som kan skapa oro och osäkerhet. Prata med andra om din oro och ta hand om din hälsa på de sätt du kan. Bra mat, sömn och träning på ett sätt som passar dig, minskar stress. Begränsa flödet av negativa nyheter om läget i världen. Hitta en nivå som är lagom för dig själv!Ge tid åt sådant som får dig att må bra. Kultur, förströelse, att skapa och få tänka på roliga saker är viktigt även när tiderna känns tunga. För vår motståndskraft och uthållighet är det viktigt att må bra, exempelvis  genom gemenskap, spel, lek, kultur, musik och idrott. Aktiviteter som väldigt många människor saknade under pandemin men som vi nu har all frihet att utöva.\nFoto:Myndigheten för civilt försvar/Melker Dahlstrand"
+}
