+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_kurser-och-forelasningar-om-hemberedskap.json
@@ -1,0 +1,55 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/kurser-och-forelasningar-om-hemberedskap/",
+  "title": "Kurser och föreläsningar om hemberedskap",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Vem utbildar i hemberedskap?"
+    },
+    {
+      "level": "h2",
+      "text": "Utbildningar för olika målgrupper"
+    },
+    {
+      "level": "h3",
+      "text": "Föreningar, organisationer och grannar"
+    },
+    {
+      "level": "h3",
+      "text": "Arbetskamrater"
+    },
+    {
+      "level": "h3",
+      "text": "Familj"
+    },
+    {
+      "level": "h3",
+      "text": "Odlingsintresserade"
+    },
+    {
+      "level": "h3",
+      "text": "Boende på landsbygd"
+    },
+    {
+      "level": "h3",
+      "text": "Personer med assistans"
+    },
+    {
+      "level": "h3",
+      "text": "Du som har sällskapsdjur"
+    },
+    {
+      "level": "h3",
+      "text": "Barn och unga"
+    },
+    {
+      "level": "h3",
+      "text": "Grupper utomhus"
+    },
+    {
+      "level": "h2",
+      "text": "Kontaktuppgifter till alla frivilliga försvarsorganisationer"
+    }
+  ],
+  "content": "Du som vill ha en föreläsning eller kurs i hemberedskap för grannskapet, skolklassen eller arbetsplatsen kan hitta den här.\nUtbudet av material om hemberedskap har verkligen vuxit. Här listas några som är aktuella 2026. Antingen frivilliga försvarsorganisationer eller andra organisationer som har löpande uppdrag från Myndigheten för civilt försvar att hålla både kortare och längre utbildningar. Det kan också vara forskningsprojekt där vi som myndighet finns med som referenspunkt eller finansiär, eller helt fristående aktörer som fack- eller studieförbund. Utbildningarna kan vara digitala eller i form av besök i er verksamhet.\nUtbildning i hemberedskap, Botkyrka kommun 2024 . Foto: Melker Dahlstrand\nTillsammans i kris och krig – beredskap för föreningar: En ny studiecirkel som utgår från broschyren ”Om krisen eller kriget kommer”. Här kan din förening diskutera broschyren utifrån ett föreningsperspektiv, göra samarbetsövningar och att skriva en beredskapsplan.Studiecirkeln är framtagen av Södra civilområdeskansliet tillsammans med studieförbunden och nätverk för civilsamhället och kommer att finnas tillgänglig via alla studieförbund och är gratis för föreningar.\nTill Beredskap för föreningar (länsstyrelsen Kronoberg)\nStudieförbundet Vuxenskolan har i samarbete med Bygdegårdarnas riksförbund tagit fram \"Bygdens beredskap\" som ska lära föreningar, organisationer och individer förutsättningar att bli en aktiv kraft i kris och krig.\nTill Bygdens beredskap på sv.se\nBoende i HSB har producerat ett eget material för grannar i samråd med oss på Myndigheten för civilt försvar. Det har delats ut till boende.\nKrisberedskap - är du förberedd: En vägledning från Medborgarskolan. Ge dig själv, familj, vänner och grannar förutsättningarna för att klara dig igenom en allvarlig händelse eller kris.\nTill Medborgarskolan\nBeredskap på arbetsplatsen: De flesta arbetsplatser ska fungera även under svåra förhållanden. Fackliga organisationer har en nyckelroll i att skapa trygghet, struktur och uthållighet. LO, TCO och Saco har tagit fram ett gemensamt material för förtroendevalda.\nTill Beredskap på arbetsplatsen på tco.seFacklig beredskap som studiecirkel: Samla minst tre personer som vill delta i en studiecirkel om beredskap och kontakta sedan ditt lokala ABF-kontor. ABF kan bidra med pedagogiska resurser, material och även lokaler eller digitala rum för era möten.\nTill Facklig beredskap på lo.se\nSköt dig själv: Stärker din förmåga att klara dig och din familjs behov i en kris. Deltagarna träffas på plats, får möta en kommunrepresentant och arbeta i grupp.\nTill Svenska Lottakåren\nDin trygghet: Vad vi själva kan göra för överlevnad i kris och för att förebygga skador och olyckor i vardagen. Anpassas av instruktören till aktuell målgrupp.\nTill Civilförsvarsförbundet\nFör din säkerhet: Grundläggande kunskaper om var man hittar information vid kriser och att stärka sin förmåga att ta hand om sig själv och sina nära.\nTill Försvarsutbildarna\nVad händer om maten tar slut i affären? En ny studieplan från Svenska Blå Stjärnan och Studiefrämjandet är ett första steg mot att stärka din och andras matberedskap genom odling.\nTill Studiefrämjandet\nBeredda tillsammans: Bygdegårdarnas Riksförbunds kurs som arrangeras i bygdegårdar runt om i landet.\nTill Bygdegårdarnas Riksförbund\nProjekt krisstöd: En digital utbildning och andra material som kan användas som stöd före, under och efter en kris till alla uppdrag inom personlig assistans.\nTill Intressegruppen för Assistansberättigade (IFA)\nStudiefrämjandet samarbetar med Svenska Blå Stjärnan och kan hjälpa till att sätta i gång med studiecirkeln ”Krisberedskap för dig och dina sällskapsdjur”.\nTill Svenska Blå Stjärnan\n\"Rädd eller beredd\" vänder sig främst till unga som är på väg ut i vuxenlivet och som kanske är på väg att flytta hemifrån. Den bygger på scenariediskussioner, värderingsövningar och deltagaraktivitet och  genomförs under två timmar tillsammans med en ledare från Civilförsvarsförbundet.\nTill Civilförsvarsförbundet\nVar beredd - inte rädd: Studieförbundet Bilda utbildar ledare och deltagare i scoutverksamhet och andra ungdomsförbund i hur man kan hantera allvarliga händelser och kriser.\nTill Bilda\nFör din säkerhet: Folk och försvar har en omfattande ungdomsverksamhet där de erbjuder kunskap, gemenskap och utmaningar inom både försvar och krisberedskap.\nTill Säkerhetspolitisk grundkurs på folkochforsvar.se\nDemokratisk beredskap: En interaktiv workshop om vilket ansvar du har när samhället drabbas av en kris och vad som menas med demokratisk beredskap.\nTill Svenska Lottakåren\nKrisvis: Friluftsliv och krisberedskap hör ihop. Det finns många likheter mellan att klara sig en vecka på fjället och en vecka hemma utan el.\nTill Friluftsfrämjandet\nDe frivilliga försvarsorganisationerna är ideella föreningar med särskilda uppdrag inom krisberedskap och totalförsvar. Idag är nästan 380 000 personer engagerade på olika sätt i de 18 organisationerna som bedriver verksamhet över hela landet. Många av dem håller kurser.\nTill frivilligforsvaret.se\nHör av dig om du vet fler aktörer och utbildningar:beredskapsveckan@mcf.se"
+}
