+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_mat.json
@@ -1,0 +1,15 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/mat/",
+  "title": "Mat",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Odla själv"
+    },
+    {
+      "level": "h2",
+      "text": "Tillgång till mat i kris"
+    }
+  ],
+  "content": "Här finns tips på vilken slags mat som är bra att ha hemma vid kris. Du kan också stärka din hemberedskap genom att odla en del av din egen mat.\nHär finns tips på vilken slags mat som är bra att ha hemma vid kris. Du kan också stärka din hemberedskap genom att odla en del av din egen mat.\nDu kan stärka din hemberedskap genom att odla en del av din egen mat.\nHär finns tips på vilken slags mat som är bra att ha hemma och hur du bäst förvarar den om elen skulle försvinna."
+}
