+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_mat_att-odla-sjalv.json
@@ -1,0 +1,19 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/mat/att-odla-sjalv/",
+  "title": "Odla själv",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Beredskapsodling"
+    },
+    {
+      "level": "h2",
+      "text": "Exempel på mat att odla"
+    },
+    {
+      "level": "h2",
+      "text": "Tips!"
+    }
+  ],
+  "content": "Du kan stärka din hemberedskap genom att odla en del av din egen mat. Om du inte har en trädgård eller kolonilott kan du odla på balkongen eller fönsterbrädan. Lär dig också något om att förvara maten.\nAtt odla, samla fröer till eget utsäde, fiska, jaga och plocka bär har varit livsviktigt under kriser och krig i historien. Det stärker också gemenskapen i bostads- och fritidsområden att vara ute och göra något tillsammans.\nFoto:Myndigheten för civilt försvar/Melker Dahlstrand\nI städerna har koloniträdgårdar haft en historisk roll för matberedskapen under kriser. Stockholms koloniförening skapade potatislotter för att stockholmarna skulle få mat under hungersnöden 1917. Även under andra världskriget var koloniträdgårdarna centrala i matförsörjningen.\nAtt odla för att ha mat i beredskap vid matbrist kallas beredskapsodling. Den ska vara enkel och ge en god skörd av energirika grödor. Idag finns inga särskilda råd från myndigheterna kring beredskapsodling men vi ser att många människor vill lära sig mer om det. Se länkarna här på sidan till organisationer du kan kontakta.\n- Potatis och andra rotsaker. Kan ligga kvar i jorden med halmtäcke fram till sen höst. Lätta att förvara efter skörd.\n- Ärtor, linser och bönor. Odla ätbara skott i fönsterbrädan. Mycket lätt att förvara länge. Torka eller frys in.\n- Lök och vitlök. Går att odla mycket enkelt från sättlök. Förvaras i rumstemperatur.\n- Kål. Bidrar med många viktiga näringsämnen. Hela plantan är ätbar.\nTa vara på frukt, bär och ätlig svamp. Ta skott från gran på våren. De är rika på C-vitamin och mineraler. Kom ihåg att allemansrätten inte gäller granskott. Du måste fråga den som äger träden om du får plocka skotten.\nOdla tillsammans\nMånga människor har redan goda kunskaper i odling och hur man tar hand om sin skörd. Vi blir starkare och tryggare genom gemenskap och att odla passar bra att göra tillsammans. Att vara utomhus och samtala kan kännas bra i oroligare tider. Inkludera även den som inte alltid har lätt att komma ut.\nLokala trädgårdsföreningar har både kunskap och utsäde/frön. Ta kontakt eller bli medlem.Informationen är framtagen av Myndigheten för civilt försvar tillsammans med Fritidsodlingens Riksorganisation, FOR"
+}
