+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_mat_tillgang-till-mat-i-kris.json
@@ -1,0 +1,43 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/mat/tillgang-till-mat-i-kris/",
+  "title": "Tillgång till mat i kris",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Du behöver mat"
+    },
+    {
+      "level": "h2",
+      "text": "Laga mat utan el"
+    },
+    {
+      "level": "h2",
+      "text": "Förvara maten rätt"
+    },
+    {
+      "level": "h2",
+      "text": "Odling"
+    },
+    {
+      "level": "h2",
+      "text": "Glöm inte kontanterna!"
+    },
+    {
+      "level": "h2",
+      "text": "Tips på matvaror"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Mat"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Preppa med liten budget"
+    },
+    {
+      "level": "h2",
+      "text": "Var kan man köpa de saker som Myndigheten för civilt försvar tipsar om?"
+    }
+  ],
+  "content": "Här finns tips på vilken slags mat som är bra att ha hemma, hur den kan tillagas och hur du bäst förvarar den vid återkommande elbrist eller ett längre strömavbrott.\nFoto:Johnér\nFoto: Johnér\nOm du är frisk kan kroppen egentligen klara sig länge utan mat, men din mentala förmåga försämras snabbt om du inte äter. Du både tänker sämre och mår sämre. Se därför till att du har något att äta! Köp extra av sådant som du ändå brukar köpa annars, som har lång hållbarhet. Det är också bra att ha tillgång till mat som inte kräver uppvärmning eller avancerad tillagning.\nVid till exempel elbrist eller ett längre strömavbrott kan det bli problem med leveranserna av varor till butikerna. Efter bara ett par dagar blir det tomt på hyllorna. Därför är det bra att ha mat hemma, så att du klarar dig minst en vecka utan att behöva handla.\nDen mesta av utrusningen i ett modernt kök slutar att fungera vid ett strömavbrott. Om det är strömavbrott kan du använda sprit- eller campingkök (stormkök) att laga mat på. Stormköket måste stå plant på ett underlag som inte kan börja brinna. Det ska också vara god ventilation, och du måste kunna släcka om det skulle bli en brand. Det allra enklaste är att vara utomhus. En utegrill är också bra för matlagning vid strömavbrott. Om du bor i lägenhet kan du gå till en grillplats i en park eller på en innergård, om det finns någon sådan nära dig.\nDet är viktigt att förvara mat och livsmedel rätt om kyl och frys inte fungerar, annars kan du bli sjuk även om du kokar eller steker maten. Börja med att äta upp det som finns i kylskåpet, medan det fortfarande är kallt och maten är färsk. Även om strömmen försvinner håller frysen sig tillräckligt kall i ett par dygn. Öppna frysdörren så få gånger som möjligt, och stäng den snabbt igen så att kylan håller sig kvar. Har du balkong eller altan kan du förvara maten där när det är kallt ute.\nDu kan stärka din hemberedskap genom att odla en del av din egen mat. Om du inte har en trädgård eller kolonilott kan du odla på balkongen eller fönsterbrädan. Lär dig också något om att förvara maten.\nMer om odling på mcf.se\nDu kan behöva ha kontanter för att kunna handla. Vid strömavbrott fungerar varken uttagsautomater eller kortbetalning men många matbutiker kan ta emot kontanter. Ta i förväg ut kontanter om kan, och förvara på ett säkert ställe. Räkna på vad du behöver för en vecka.\nFörbered dig så du inte måste skaffa allt på en gång om något allvarligt händer. Anpassa råden i checklistorna efter dina behov och förutsättningar. Vissa saker kanske du kan dela med andra, exempelvis dina grannar. I kris och krig måste vi alla hjälpa varandra, men också i vardagen."
+}
