+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_medicin-for-en-manad-hemma.json
@@ -1,0 +1,15 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/medicin-for-en-manad-hemma/",
+  "title": "Medicin för en månad hemma",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "För vem gäller rekommendationen?"
+    },
+    {
+      "level": "h2",
+      "text": "Ingår inte i rekommendationen"
+    }
+  ],
+  "content": "Om du brukar få medicin eller förbrukningsartiklar utskrivna av hälso- och sjukvården är det bra att ha hemma så att det räcker för en hel månad. Då klarar du dig bättre om det uppstår en kris eller en bristsituation.\nGenerellt för hemberedskap är rekommendationen en veckas förråd av till exempel mat. Men när det gäller receptbelagda mediciner för långvarig behandling är det bra att ha för en hel månad hemma. Det är Socialstyrelsens rekommendation till dig som behandlas under en längre tid.\nPrecis som i ditt skafferi behöver du tänka på att ha en rutin för att rotera lagret så att du använder de äldsta medicinerna först.\nMånga privatpersoner har redan idag en god beredskap med sin medicin, eftersom det även tidigare varit möjligt att hämta ut i god tid. Så om du redan har en månads förbrukning hemma ska du inte hämta ut mer.\nRekommendationen gäller dig som under längre tid behandlas med receptbelagda mediciner, till exempel medicin som används vid:\n- Förhöjt blodtryck\n- Hjärtåkomma\n- Astma, KOL och andra lungsjukdomar\n- Mag- och tarmsjukdomar\n- Njursjukdomar\n- Diabetes\nRekommendationen gäller också dig som under en längre tid använder förbrukningsartiklar som du fått utskrivna av hälso- och sjukvården, till exempel utrustning för kontroll av blodsockret.\n- Receptfria mediciner och förbrukningsartiklar.\n- Mediciner som skrivs ut för en kortare tids behandling.\n- Mediciner eller dosering som provas ut och kan komma att ändras.\n- Mediciner som av medicinska skäl är utskrivna så att det inte går att hämta ut nästa förpackning förrän en viss tid har passerat.\n- Mediciner som är färdigförpackade i en påse för varje dos och ibland med flera mediciner som ska tas samtidigt.\n- Livsmedel som du har fått utskrivna på recept.\nEn del mediciner är utskrivna så att det inte går att hämta ut en månad i förväg och då gäller inte rekommendationen. Det gäller till exempel vissa narkotikaklassade mediciner.\nI vissa fall kan det också vara svårt att ordna en månads beredskap hemma. Varorna kan vara tillfälligt slut eller svåra att lagra hemma, till exempel om de tar mycket plats, behöver förvaras i kylskåp eller är brandfarliga. I de fallen är det bättre att ha en så god beredskap som möjligt än ingen alls.\nRekommendationen gäller på lång sikt. Det finns ingen generell brist på mediciner eller förbrukningsartiklar i Sverige men det är viktigt att stärka sin hemberedskap för att vi alla i samhället ska vara bättre förberedda om det uppstår bristsituationer."
+}
