+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_tillgang-till-kommunikation-i-kris.json
@@ -1,0 +1,59 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/tillgang-till-kommunikation-i-kris/",
+  "title": "Tillgång till kommunikation i kris",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Lyssna på radio"
+    },
+    {
+      "level": "h2",
+      "text": "Om du inte kan ringa 112"
+    },
+    {
+      "level": "h2",
+      "text": "Krisinformation.se samlar viktig information"
+    },
+    {
+      "level": "h2",
+      "text": "Saklig information"
+    },
+    {
+      "level": "h2",
+      "text": "Ring informationsnumret 113 13"
+    },
+    {
+      "level": "h2",
+      "text": "Kommunikation utan el"
+    },
+    {
+      "level": "h2",
+      "text": "Skriv ner viktiga nummer på papper"
+    },
+    {
+      "level": "h2",
+      "text": "Kommunens trygghetspunkter vid kris"
+    },
+    {
+      "level": "h2",
+      "text": "Ladda elprylarna i förväg"
+    },
+    {
+      "level": "h2",
+      "text": "Hjälp varandra!"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Kommunikation"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Preppa med liten budget"
+    },
+    {
+      "level": "h2",
+      "text": "Var kan man köpa de saker som Myndigheten för civilt försvar tipsar om?"
+    }
+  ],
+  "content": "Vid en större olycka eller en kris behöver du få veta vad som händer, hur du själv ska agera och vad ansvariga myndigheter gör. Du kan också behöva komma i kontakt med dina anhöriga.\nSveriges Radio P4 är beredskapskanal och fortsätter att sända vid kris och krig. Där får du veta viktiga detaljer om hur läget är i samhället, lokala väderprognoser, var du kan hämta vatten, vart du kan bege dig om du behöver hjälp eller om det är evakuering på gång. För att kunna höra radion även om det är strömavbrott eller om din mobiltelefon inte har internet, kan du använda något av det här:\n- En radio som går på batteri\n- En radio som kan laddas med solceller\n- En radio som går att veva igång\n- Bilradio\nOm du akut behöver ringa 112 och telefonen inte funkar, kan du bege dig till närmaste brandstation och låna en nödtelefon där.\nFör att det ska vara enkelt att hitta tillförlitlig och uppdaterad information dygnet runt vid en allvarlig händelse finns det en särskild webbplats från de myndigheter som arbetar med det som har hänt: Krisinformation.se. Här finns all myndighetsinformation samlad på ett ställe, så att du själv slipper hålla reda på vilken myndighet som ansvarar för vad och hur du kommer i kontakt med rätt myndighet. Myndigheten för civilt försvar ansvarar för webbplatsen.\nViktig lokal information kan finnas på kommunens webbplats och kommunens sociala medier. Ditt elbolag kan också ofta berätta om hur lång tid det beräknas ta innan strömmen är tillbaka.\nDet är viktigt att människor litar på den information som vi myndigheter förmedlar. Det betyder att informationen måste vara korrekt och begriplig så att den inte skapar förvirring och misstro hos människor. Vikten av snabb och pålitlig information vid olyckor och kriser uppmärksammas allt mer. Idag arbetar ansvariga myndigheter aktivt för att bli bra på att informera allmänheten vid svåra händelser.\n113 13 är Sveriges nationella informationsnummer. Dit kan du ringa om du vill få information vid allvarliga olyckor och kriser i samhället. Det kan till exempel handla om en större storm, en influensaepidemi, en stor trafikolycka eller en översvämning. All information som lämnas på 113 13 är verifierad, det vill säga avstämd med till exempel kommunen där händelsen har ägt rum.\nRing 113 13 när:\n- du vill få information om en pågående kris i samhället\n- du själv vill lämna information om sådana större händelser i samhället.\nAlla kommuner har en skyldighet att kunna informera alla som bor, verkar eller vistas i kommunen vid en allvarlig händelse, och ska snabbt kunna lägga ut information på kommunens webbplats. Men vid ett elavbrott blir arbetet svårare. Kommunen kan då använda sig av allt från flygblad till torgmöten och personliga informatörer för att nå ut med viktig information.\nOm din mobil inte fungerar kommer du inte heller åt dina kontakter där. Kan du numret till anhöriga, skolan, kommunen och elleverantören? Din fasta telefon, eller någon annans telefon kanske fortfarande fungerar. Ha dina viktigaste telefonnummer uppskrivna på en papperslapp och lägg den i krislådan eller sätt upp på en plats som alla i hushållet känner till.\nI en del kommuner finns trygghetspunkter som öppnas om det blir en kris. Det kan vara i en skola, en gymnastiksal, ett församlingshem eller i en kyrka. På trygghetspunkten kan det exempelvis finnas information, värme, dricksvatten och möjlighet till att ladda mobiltelefonen. Du kan i förväg ta reda på om din kommun har trygghetspunkter och var dessa finns.\nOm du hör talas om att det ska bli oväder – passa på att i förväg ladda allt som går. En powerbank eller annat slags extrabatteri till mobilen kan vara bra att ha.\nHåll kontakt med grannar och anhöriga på de sätt du kan, och hör efter om de behöver hjälp."
+}
