+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_tillgang-till-varme-i-kris.json
@@ -1,0 +1,43 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/tillgang-till-varme-i-kris/",
+  "title": "Tillgång till värme i kris",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Täta golv, dörrar och fönster"
+    },
+    {
+      "level": "h2",
+      "text": "Ta fram vinterkläderna"
+    },
+    {
+      "level": "h2",
+      "text": "Alternativ värme"
+    },
+    {
+      "level": "h2",
+      "text": "Kom ihåg!"
+    },
+    {
+      "level": "h2",
+      "text": "Undvik frysta vattenledningar"
+    },
+    {
+      "level": "h2",
+      "text": "Tänk på säkerheten!"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Värme i kris"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Preppa med liten budget"
+    },
+    {
+      "level": "h2",
+      "text": "Var kan man köpa de saker som Myndigheten för civilt försvar tipsar om?"
+    }
+  ],
+  "content": "Din bostad blir snabbt kall om det är strömavbrott på vintern. Att vistas i en bostad utan värme är en stor påfrestning, men det finns mycket du kan göra för att situationen ska bli lättare.\nDet som blir kallt först är golven. Lägg på extra mattor eller filtar. Täta fönster- och dörrspringor med tejp, filtar eller plast för att spara på värmen. Men glöm inte att utrymning måste kunna ske på ett säkert sätt.\nVälj ett rum att vara i. Häng filtar för fönstren och täck golvet med mattor.\nTänk som du skulle göra inför en kall vinterdag. Använd liggunderlag, skidkläder, mössa och vantar. Klä dig lager på lager. Då blir det luft mellan plaggen som värms upp av kroppen. Ullplagg ger bra värme och ska vara närmast kroppen. Det är lättare att hålla värmen om ni är flera i samma rum. Bygg en koja att sova i, till exempel under ett bord som täcks med filtar. Ta på varma tröjor, sockar och mössa och kryp ner i varsin sovsäck.\nOm du har en fungerande kakelugn, öppen spis eller kamin är de till stor hjälp. Andra exempel på värmekällor är fotogen- eller gasolkaminer. För att kunna använda utrustningen på ett säkert sätt är det viktigt att följa anvisningarna. Det finns till exempel regler för:\n- placering\n- avstånd\n- ventilation\nStearin- och värmeljus ger både ljus och lite värme. Men se upp med brandfaran. Använd ljusstakar eller ställ ljusen i ett fat med vatten. Ställ alltid värmeljus glest! Glöm inte att ha extra tändstickor eller braständare.\nDet är mycket viktigt att vädra regelbundet för att släppa in syre!\nVid strömavbrott finns det risk för att vattnet fryser i rör och ledningar. Det kan leda till vattenskador. För att förebygga att vattnet fryser kan rören isoleras ordentligt. Om du misstänker att en ispropp har bildats i dina rör, försök absolut inte att tina den med öppen låga. Det kan vara mycket brandfarligt. Stäng i stället av vattnet och kontakta en VVS-firma och ditt försäkringsbolag.\nSå fort vardagen inte fungerar som vanligt är det lätt att falla ur sina säkerhetsvanor. Här är några säkerhetstips vid strömavbrott:\n- Elda på ett säkert sätt genom att följa anvisningarna för den fasta eller tillfälliga uppvärmningsanordningen.\n- Se till att det finns ventilation.\n- Använd inte kamin eller levande ljus när du sover.\n- Var försiktig när du hanterar bränsle och öppna lågor inomhus.\n- Placera levande ljus långt bort från brännbara material, till exempel gardiner.\n- Ställ värmeljus glest i en ljusstake som inte kan börja brinna.\n- Placera och använd stormkök eller campingkök på en säker plats, tillräckligt långt från exempelvis skåp och köksfläkt. Helst utomhus."
+}
