+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_vatten.json
@@ -1,0 +1,15 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/vatten/",
+  "title": "Vatten",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "Tillgång till vatten i kris"
+    },
+    {
+      "level": "h2",
+      "text": "Tvätt och toalett"
+    }
+  ],
+  "content": "Brist på vatten är mycket allvarligare än brist på mat. Därför är det viktigt att ha en plan för att kunna ordna vatten i alla lägen, till både dryck och matlagning. Här finns också tips om hygien och toalett.\nBrist på vatten är mycket allvarligare än brist på mat. Därför är det viktigt att ha en plan för att kunna ordna vatten i alla lägen, till både dryck och matlagning. Här finns också tips om hygien och toalett.\nBrist på vatten är mycket allvarligare än brist på mat. Därför är det viktigt att ha en plan för att kunna ordna vatten.\nTips för hur du kan gå på toaletten och tvätta dig själv och dina kläder utan vatten eller elektricitet."
+}
