+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_vatten_hygien-tvatt-och-toalett.json
@@ -1,0 +1,31 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/vatten/hygien-tvatt-och-toalett/",
+  "title": "Tvätt och toalett i kris",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "På den här sidan"
+    },
+    {
+      "level": "h2",
+      "text": "Toalettbesök utan vatten"
+    },
+    {
+      "level": "h3",
+      "text": "Kommunen ansvarar för avfallet"
+    },
+    {
+      "level": "h2",
+      "text": "Personlig hygien"
+    },
+    {
+      "level": "h2",
+      "text": "Tvätta kläder utan el"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Hygien"
+    }
+  ],
+  "content": "Tips för hur du kan gå på toaletten och tvätta dig själv och dina kläder utan vatten eller elektricitet.\nOm det inte finns vatten på grund av strömavbrott eller något annat, kan du inte spola i toaletten. Det kan snabbt bli stopp i avloppet om du ändå fortsätter att använda toaletten, speciellt i flerbostadshus.\nBeroende på hur du bor, kan det förstås vara enklast att göra dina behov utomhus. Då är det viktigt att visa hänsyn enligt Allemansrätten. Välj en plats långt ifrån vattendrag och där folk går. Gräv en grop först. Gör dina behov. Täck sedan över rejält, med jord eller en stor sten. Utedass och husvagnar är också ofta användbara i kris!När du ska kissa\nKissa i toaletten men lägg pappret i en hink eller papperskorg bredvid. Lägg på ett lock som sluter tätt så sprider sig inte lukten i resten av bostaden. Ta reda på var du kan slänga avfallet och var noga med handhygienen.\nNär du ska bajsa\nDet kan kännas enklast att sitta på toaletten som vanligt, med en plastpåse eller sopsäck i toalettstolen. Använd en biologiskt nedbrytbar påse. När du är klar med toalettbesöket, strö över till exempel kattsand, kompostströ eller sågspån. Knyt ihop påsen. Din kommun informerar om hur du ska slänga avfallet. Var noga med handhygienen.\nDet finns också mer avancerade toaletter som fungerar utan vatten och el:\n- Förbränningstoaletter som bryter ner bajset till aska. De drivs med gasol.\n- Ihopfällbara toaletter med urinseparerande funktion. Toaletten ser ut som en vanlig campingstol med en påse undertill. Bajset hamnar i påsen och kisset drivs ut med hjälp av en slang.\nRåden på den här sidan är allmänna. Varje kommun ska sedan ha en plan för att ta hand om avfallet. De planerna kan se olika ut. Här är exempel:\n- Kommunen placerar ut särskilda containers eller flak där man kan kasta avfallet i påsar\n- Kommunen ställer ut portabla toaletter av typen som brukar användas på till exempel festivaler och badplatser (Bajamajor)\n- Kommunen tar hand om hämtning av latrintunnor från hushållen. Som privatperson behöver du då skaffa en sådan tunna.\nKontakta din kommun för att veta vad som gäller!\nOm det blir brist på vatten är det viktigast att du använder det vatten du har till att dricka och laga mat. Du kommer antagligen inte att kunna bada, duscha eller tvätta dig som du är van vid. Här finns några tips på sätt att tvätta sig som sparar mycket vatten eller inte kräver vatten alls.Fempunktstvätt\nFör att tvätta dig själv eller hjälpa någon annan att tvätta sig, räcker det ofta med en så kallad fempunktstvätt. Till det behövs bara lite vatten, eller inget vatten alls. Fyll handfatet med vatten, helst uppvärmt. Tvätta med tvål och tvättlapp, från den renaste delen på kroppen och avsluta med den smutsigaste. Till exempel:\n- Ansiktet\n- Händerna\n- Armhålorna\n- Intimområdet\n- Fötterna\nOm det inte finns vatten alls, kan du använda våtservetter. Ett annat alternativ är tvättkräm som du inte behöver skölja bort. Den fungerar för både intimhygien och babyvård och finns på apotek.\nDet finns flera typer av enklare duschar som du själv fyller på vatten i och som sparar vatten. De går ofta att skaffa i friluftsbutiker, båtbutiker eller liknande.\n- Campingdusch med en vattensäck och ett munstycke. Fyll säcken med vatten, häng upp den och låt vattnet rinna genom munstycket med självtryck.\n- Portabla duschmunstycken består av pump, slang och munstycke. Duschen pumpar själv upp vatten från en behållare och laddas med USB.\n- Genomströmningsdusch kopplas till en vattenbehållare och värmer upp vattnet med hjälp av gasol.\n- Du kan också bygga en enkel dusch av en vanlig hink med ett eller flera hål i botten. Häng upp hinken utomhus och låt vattnet strömma genom hålet eller hålen.\nUtan el fungerar inte tvättmaskinen. Om du har tillgång till vatten kan du tvätta dina kläder på andra sätt.\nLägg kläderna i en behållare. Värm upp vattnet och häll det i behållaren. Tillsätt tvättmedel, helst handtvättmedel eller tvättvål. Låt kläderna ligga i blöt ett tag. Rör runt kläderna då och då och tvätta mer intensivt mot slutet. Skölj väl efteråt.\nDet finns flera olika behållare att tvätta dina kläder i. Här listas några förslag.\n- I en hink med lock, eller tät kylväska. Kläderna tvättas när du rör runt i blandningen.\n- I en tvättpåse (washbag). Kläderna tvättas när du trycker med händerna på den vattentäta säcken.\n- I en manuell tvättmaskin med vev. Kläderna tvättas när du vevar."
+}
