+++ webscraper/data/sv_rad-till-privatpersoner_hemberedskap---preppa-for-en-vecka_vatten_tillgang-till-vatten-i-kris.json
@@ -1,0 +1,51 @@
+{
+  "url": "https://www.mcf.se/sv/rad-till-privatpersoner/hemberedskap---preppa-for-en-vecka/vatten/tillgang-till-vatten-i-kris/",
+  "title": "Tillgång till vatten i kris",
+  "headings": [
+    {
+      "level": "h2",
+      "text": "På den här sidan"
+    },
+    {
+      "level": "h2",
+      "text": "Dricksvatten"
+    },
+    {
+      "level": "h3",
+      "text": "Tappa upp dricksvatten"
+    },
+    {
+      "level": "h3",
+      "text": "Dricksvatten i dunk"
+    },
+    {
+      "level": "h3",
+      "text": "Förvara vattnet mörkt och svalt"
+    },
+    {
+      "level": "h2",
+      "text": "Hämta vatten"
+    },
+    {
+      "level": "h2",
+      "text": "Koka vatten"
+    },
+    {
+      "level": "h2",
+      "text": "Du som har egen brunn"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Vatten"
+    },
+    {
+      "level": "h2",
+      "text": "Checklista: Preppa med liten budget"
+    },
+    {
+      "level": "h2",
+      "text": "Var kan man köpa de saker som Myndigheten för civilt försvar tipsar om?"
+    }
+  ],
+  "content": "Brist på vatten är mycket allvarligare än brist på mat. Därför är det viktigt att ha en plan för att kunna ordna vatten i alla lägen, till både dryck och matlagning.\nI något läge där samhällets service inte fungerar som vanligt, till exempel om det plötsligt inte kommer vatten ur kranen, måste varje hushåll ta ett ansvar för att lösa olika situationer. Det kan innebära att de rutiner man har hemma som man är van vid, behöver förändras under en period.\nMed PET-flaskor, en vanlig frys och möjlighet att koka ditt vatten kommer du långt. Räkna med att en vuxen person behöver tre till fem liter vatten per dygn. Det ska räcka till både dryck, matlagning och grundläggande hygien. Har du husdjur går det åt vatten även för dem.Det är förstås svårt att lagra stora mängder vatten hemma, men lite är bättre än ingenting. Ett extra förråd av dricksvatten kan bestå av dunkar eller andra behållare som du själv fyller med dricksvatten. Det kan även bestå av vatten på flaska som du köper i butiken.\nOm strömmen försvinner kan vattenpumparna sluta att fungera efter ett tag. Då behöver du ha extra vatten upptappat, till både familjen och husdjuren. Ett smart sätt att förbereda sig är att frysa in vatten i PET-flaskor. Flaskorna fungerar då både som reservdricksvatten och som kylklampar i frysen under längre strömavbrott. Kom ihåg att inte fylla flaskorna ända upp, då spricker de när vattnet fryser.\nAnvänd rena och väl förslutna dunkar som är avsedda att användas för dricksvatten. Fyll på med dricksvatten av bra kvalitet, till exempel kommunalt dricksvatten. Förvara vattnet mörkt och så svalt som möjligt.\nRåd om rengöring och övrigt hantering av dunkar och finns hos Livsmedelsverket.\nDet går inte att säga hur länge vatten i dunk kan förvaras. Det beror på vilka förhållanden som råder och du får själv avgöra om du vill dricka vattnet. Det kan du göra genom att titta, lukta och smaka på vattnet. Under riktigt bra förhållanden kan dricksvatten i dunk hålla sig fräscht i många månader.\nTa gärna för vana att regelbundet kontrollera om vattnet är ok en eller ett par gånger per år. Byt vatten i ditt dricksvattenförråd om det behövs, exempelvis genom att använda det till bevattning.\nOm det blir brist på vatten kan kommunen ställa ut vattentankar på bestämda platser, dit du får gå för att hämta vatten i egna kärl. Räkna med att du får hämta begränsade mängder, eftersom vattnet ska räcka till så många som möjligt. I nödläge kan du hämta vatten från vattendrag, sjöar och dammar. Ta i så fall vatten från ställen där vattnet rinner, eller som ligger långt ut från stranden. Koka alltid vatten som du har hämtat ute.\nI krissituationer är förorenat vatten en av de vanligaste smittspridarna av mag- och tarmsjukdomar. Om det finns misstanke om att kranvattnet innehåller bakterier, virus eller parasiter kan du behöva rena det. Detsamma gäller för vatten du har hämtat ute, eller om du använder snö och is. Det ska kokas innan du kan dricka det eller använda det i matlagning.\n- Koka upp vattnet tills det bubblar kraftigt med stora bubblor. Bakterier, virus och parasiter dör när vattnet kokas.\n- Häll upp det kokta dricksvattnet i en ren kanna eller flaska, och låt det svalna.\nDet finns också andra vattenreningsmetoder, till exempel tabletter och filter. De metoderna kan rena vattnet från skadliga mikroorganismer, men hjälper inte mot kemiska föroreningar.\nEn egen dricksvattenanläggning, till exempel en brunn, påverkas inte av störningar i den kommunala vattenförsörjningen. Du som fastighetsägare ansvarar själv för vattnet.\nGör en plan för hur du får upp vattnet om elen inte fungerar. Det kan till exempel vara med hjälp av\n- en hink som du hissar ner i brunnen\n- en handpump\n- ett reserv-elverk.\nOlika krishändelser kan påverka kvaliteten på vattnet och möjligheterna att ta upp vatten. Till exempel\n- elavbrott\n- olyckor\n- miljöföroreningar\n- radioaktivt nedfall\n- väder och naturhändelser.\nEfter en sådan händelse ska du alltid ta ett vattenprov och skicka in för analys. Läs mer hos Livsmedelsverket.\nKom ihåg att du som har egen brunn själv ansvarar för att vattnet är säkert, och att det räcker. Gör kontroller regelbundet."
+}
